+++ api/posts.js
@@ -1,6 +1,8 @@
 const router = require('express').Router();
 
 const { validateAgainstSchema } = require('../lib/validation');
+const { generateAuthToken, requireAuthentication} = require('../lib/auth');
+const mysqlPool = require('../lib/mysqlPool');
 const {
     PostSchema,
     getPostsPage,
@@ -98,12 +100,14 @@ router.get('/:id', async (req, res, next) => {
 /*
  * Route to replace data for a post.
  */
-router.put('/:id', async (req, res, next) => {
-  const id = parseInt(req.params.id);
-  if (id === req.body.postId){
-    if (validateAgainstSchema(req.body, PostSchema)) {
+router.put('/:id', requireAuthentication, async (req, res, next) => {
+  if (req.user != req.body.userId) {
+    res.status(403).send({
+      error: "Unauthorized to access the specified resource"
+    });
+  } else {
       try {
-        
+        const id = parseInt(req.params.id)
         const updateSuccessful = await replacePostById(id, req.body);
         if (updateSuccessful) {
           res.status(200).send({
@@ -120,16 +124,6 @@ router.put('/:id', async (req, res, next) => {
           error: "Unable to update specified post.  Please try again later."
         });
       }
-    } else {
-      res.status(400).send({
-        error: "Request body is not a valid post object"
-      });
-    }
-  }
-  else{
-    res.status(500).send({
-      error: "Unable to update specified post. Can not change the post id."
-    });
   }
 });
 
@@ -137,28 +131,30 @@ router.put('/:id', async (req, res, next) => {
 /*
  * Route to delete a post.
  */
-router.delete('/:id', async (req, res, next) => {
-    try {
-      const deleteSuccessful = await deletePostById (parseInt(req.params.id));
-      if (deleteSuccessful) {
-        res.status(204).end();
-      } else {
-        next();
+router.delete('/:id', requireAuthentication, async (req, res, next) => {
+  const [ results ] = await mysqlPool.query(
+    'SELECT * FROM posts WHERE id = ?',
+    [ req.params.id]
+  );
+  if (req.user != results[0].userId) {
+    res.status(403).send({
+      error: "Unauthorized to access the specified resource"
+    });
+  } else {
+      try {
+        const deleteSuccessful = await deletePostById (parseInt(req.params.id));
+        if (deleteSuccessful) {
+          res.status(204).end();
+        } else {
+          next();
+        }
+      } catch (err) {
+        console.error(err);
+        res.status(500).send({
+          error: "Unable to delete post.  Please try again later."
+        });
       }
-    } catch (err) {
-      console.error(err);
-      res.status(500).send({
-        error: "Unable to delete post.  Please try again later."
-      });
     }
   });
-
-
-
-
-
-
-
-
 
 exports.router = router;
